+++ src/saving_loading/texture_init.c
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/27 20:18:50 by freddy            #+#    #+#             */
-/*   Updated: 2024/07/05 18:54:22 by freddy           ###   ########.fr       */
+/*   Updated: 2024/07/05 23:59:10 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -106,6 +106,20 @@ static inline void	load_elements_overview_textures(void)
 	game()->textures.hightlight_earth = l("hud/elements_overview/earth.png");
 }
 
+static inline void	load_explosion(void)
+{
+	game()->textures.explosion_1 = l("entities/explosion/1.png");
+	game()->textures.explosion_2 = l("entities/explosion/2.png");
+	game()->textures.explosion_3 = l("entities/explosion/3.png");
+	game()->textures.explosion_4 = l("entities/explosion/4.png");
+	game()->textures.explosion_5 = l("entities/explosion/5.png");
+	game()->textures.explosion_6 = l("entities/explosion/6.png");
+	game()->textures.explosion_7 = l("entities/explosion/7.png");
+	game()->textures.explosion_8 = l("entities/explosion/8.png");
+	game()->textures.explosion_9 = l("entities/explosion/9.png");
+	game()->textures.explosion_10 = l("entities/explosion/10.png");
+}
+
 void	load_static_textures(void)
 {
 	game()->textures.door = l("entities/door.png");
@@ -131,6 +145,7 @@ void	load_static_textures(void)
 	load_status_animatons();
 	load_blight_textures();
 	load_elements_overview_textures();
+	load_explosion();
 	game()->textures.elementor_idle = l("entities/elementor/opacity.png");
 	game()->textures.elementor_hurt = l("entities/elementor/hurt.png");
 }